backtester/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dnspython==2.8.0
idna==3.11
numpy==2.2.6
pandas==2.3.3
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pyarrow==22.0.0
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tqdm==4.67.1
tzdata==2025.3
urllib3==2.6.2
pydantic>=2.0.0
tenacity>=8.2.0